Fix bugs, remove dead code and harden build flags
- Fix warn/critical logic in sslcheck (WARNING threshold was unreachable)
- Remove dead size()==0 check inside aborted-job dedup loop (out-of-bounds risk)
- Replace int with size_t for all string::find() results, compare against string::npos
- Fix curl resource leak: call cleanup before exit() on CURL error
- Fix stray quote in javashow SOAP XML (malformed request)
- Fix typo TESTCOPMANY -> TESTCOMPANY in XMI logon
- Reduce SAP_UC stack buffers from 99999 to safe sizes (rfc_ping, ssfp_get_pseinfo)
- Remove unused variables: resultLen, nth_backslash, item_gefunden, status_* sentinels
- Remove unused time_converter.cpp and its Makefile rule
- Add const to checkConnection errInfo parameter
- Scope object_gefunden to GetAlertsResponse block
- Harden Makefile: -fPIE, -fstack-protector-strong, -D_FORTIFY_SOURCE=2,
  -Wformat=2, -Werror=format-security, -fwrapv, -fno-delete-null-pointer-checks,
  full RELRO (-z relro -z now), noexecstack; remove -Wno-unused-variable/-sign-compare

+++ read_cert_infos.cpp
@@ -42,7 +42,7 @@ vector<string> read_cert_infos(const vector<string>& certlist_array, bool ssl_ch
     vector<string> certlist_subj_valid_until;
 
     for (size_t i = 0; i < certlist_array.size(); ++i) {
-        int finde_semikolon = certlist_array[i].find(";;;");
+        size_t finde_semikolon = certlist_array[i].find(";;;");
         string cert_hex     = certlist_array[i].substr(0, finde_semikolon);
 
         int bytes   = cert_hex.size() / 3;
